+++ isitfit/cli/cost.py
@@ -5,24 +5,31 @@ import click
 
 # Use "cls" to use the IsitfitCommand class to show the footer
 # https://github.com/pallets/click/blob/8df9a6b2847b23de5c65dcb16f715a7691c60743/click/decorators.py#L92
-from isitfit.cli.click_descendents import IsitfitCommand, isitfit_group, isitfit_option_profile
+from isitfit.cli.click_descendents import IsitfitCommand, isitfit_group, isitfit_option_base, isitfit_option_profile
 
 
-# Note about below --ndays: Cannot put it in the `cost` function
-# because of a bug in click whereby `isitfit cost analyze --help`
-# would prompt for the "ndays" value, eventhough a --help was passed
-# Hence, duplicating this in `analyze` and `optimize`
-# Reference: https://github.com/pallets/click/issues/295
 @isitfit_group(help="Evaluate AWS EC2 costs", invoke_without_command=False)
 @click.option('--filter-region', default=None, help='specify a single region against which to run cost analysis/optimization')
-# @click.option('--ndays', default=90, prompt='Number of days to lookback (use `isitfit cost --ndays=90 ...` to skip this prompt)', help='number of days to look back in the data history', type=int)
 @isitfit_option_profile()
+@isitfit_option_base(
+  '--ndays',
+  default=7,
+  prompt='Number of days to lookback (between 1 and 90, use `isitfit cost --ndays=7 ...` to skip this prompt)',
+  help='number of days to look back in the data history',
+  type=click.IntRange(1, 90)
+)
 @click.pass_context
-def cost(ctx, filter_region, profile):
+def cost(ctx, filter_region, ndays, profile):
+  # FIXME click bug: `isitfit command subcommand --help` is calling the code in here. Workaround is to check --help and skip the whole section
+  import sys
+  if '--help' in sys.argv: return
+
   # gather anonymous usage statistics
   from isitfit.utils import ping_matomo
-  ping_matomo("/cost?filter_region=%s"%filter_region)
+  ping_matomo("/cost?filter_region=%s&ndays=%i"%(filter_region, ndays))
 
+  # save to click context
+  ctx.obj['ndays'] = ndays
   ctx.obj['filter_region'] = filter_region
 
   pass
@@ -32,23 +39,15 @@ def cost(ctx, filter_region, profile):
 # Check note above about ndays
 # Another note about ndays: using click.IntRange for validation. Ref: https://click.palletsprojects.com/en/7.x/options/?highlight=prompt#range-options
 @cost.command(help='Analyze AWS EC2 cost', cls=IsitfitCommand)
-@click.option(
-  '--ndays',
-  default=7,
-  prompt='Number of days to lookback (between 1 and 90, use `isitfit cost analyze --ndays=7` to skip this prompt)',
-  help='number of days to look back in the data history',
-  type=click.IntRange(1, 90)
-)
 @click.option('--filter-tags', default=None, help='filter instances for only those carrying this value in the tag name or value')
 @click.option('--save-details', is_flag=True, help='Save details behind calculations to CSV files')
 @click.pass_context
-def analyze(ctx, ndays, filter_tags, save_details):
+def analyze(ctx, filter_tags, save_details):
     # gather anonymous usage statistics
     from isitfit.utils import ping_matomo, b2l
-    ping_matomo("/cost/analyze?ndays=%i&filter_tags=%s&save_details=%s"%(ndays, filter_tags, b2l(save_details) ))
+    ping_matomo("/cost/analyze?filter_tags=%s&save_details=%s"%(filter_tags, b2l(save_details) ))
 
     # save to click context
-    ctx.obj['ndays'] = ndays
     share_email = ctx.obj.get('share_email', [])
 
     #logger.info("Is it fit?")
@@ -71,25 +70,16 @@ def analyze(ctx, ndays, filter_tags, save_details):
     mm_all.get_ifi(tqdml2)
 
 
-# check note above about ndays
 @cost.command(help='Generate recommendations of optimal EC2 sizes', cls=IsitfitCommand)
-@click.option(
-  '--ndays',
-  default=7,
-  prompt='Number of days to lookback (between 1 and 90, use `isitfit cost optimize --ndays=7` to skip this prompt)',
-  help='number of days to look back in the data history',
-  type=click.IntRange(1, 90)
-)
 @click.option('--n', default=-1, help='number of underused ec2 optimizations to find before stopping. Skip to get all optimizations')
 @click.option('--filter-tags', default=None, help='filter instances for only those carrying this value in the tag name or value')
 @click.pass_context
-def optimize(ctx, ndays, n, filter_tags):
+def optimize(ctx, n, filter_tags):
     # gather anonymous usage statistics
     from isitfit.utils import ping_matomo
-    ping_matomo("/cost/optimize?ndays=%i&n=%i&filter_tags=%s"%(ndays, n, filter_tags ))
+    ping_matomo("/cost/optimize?n=%i&filter_tags=%s"%(n, filter_tags ))
 
     # save to context
-    ctx.obj['ndays'] = ndays
     share_email = ctx.obj.get('share_email', [])
 
     #logger.info("Is it fit?")