+++ isitfit/tests/cost/test_functional.sh
@@ -40,48 +40,48 @@ time isitfit version
 
 
 echo "Test 1: on profile shadiakiki1986@gmail.com@amazonaws.com (expect in AWS_DEFAULT_REGION=us-west-2)"
-AWS_PROFILE=shadi_shadi isitfit cost analyze --ndays=90
+AWS_PROFILE=shadi_shadi isitfit cost --ndays=90 analyze
 
 
 echo "Test 2: on profile shadi@autofitcloud.com@amazonaws.com (expect in AWS_DEFAULT_REGION=eu-central-1)"
-AWS_PROFILE=afc_shadi_useast1 isitfit cost analyze --ndays=90
+AWS_PROFILE=afc_shadi_useast1 isitfit cost --ndays=90 analyze
 
 
 echo "Test 3: default profile in region with 0 ec2 instances"
 # Note, unlike isitfit tags dump which returns a non-0 code if 0 ec2 found, this one just returns 0
-isitfit cost --filter-region=eu-central-1 analyze --ndays=90 || echo "Expect to fail"
+isitfit cost --filter-region=eu-central-1 --ndays=90 analyze || echo "Expect to fail"
 
 
 echo "Test 4: optimize with default profile"
-isitfit cost optimize --ndays=90
+isitfit cost --ndays=90 optimize
 
 
 echo "Test 5: optimize in region with 0 ec2 instances"
 # Note, unlike isitfit tags dump which returns a non-0 code if 0 ec2 found, this one just returns 0
-isitfit cost --filter-region=eu-central-1 optimize --ndays=90
+isitfit cost --filter-region=eu-central-1 --ndays=90 optimize
 
 
 echo "Test 6a: optimize with n=1 on shadi@autofitcloud.com@amazonaws.com"
-AWS_PROFILE=default isitfit cost optimize --ndays=90 --n=1
+AWS_PROFILE=default isitfit cost --ndays=90 optimize --n=1
 
 echo "Test 6b: optimize with n=1 on shadiakiki1986@gmail.com@amazonaws.com"
-AWS_PROFILE=shadi_shadi isitfit cost optimize --ndays=90 --n=1
+AWS_PROFILE=shadi_shadi isitfit cost --ndays=90 optimize --n=1
 
 
 echo "Test 7: {analyse,optimize} filter-tags {ffa,inexistant}"
-isitfit cost optimize --ndays=90 --filter-tags=ffa
-isitfit cost analyze  --ndays=90 --filter-tags=ffa
+isitfit cost --ndays=90 optimize --filter-tags=ffa
+isitfit cost --ndays=90 analyze  --filter-tags=ffa
 
-isitfit cost optimize --ndays=90 --filter-tags=inexistant
-isitfit cost analyze  --ndays=90 --filter-tags=inexistant
+isitfit cost --ndays=90 optimize --filter-tags=inexistant
+isitfit cost --ndays=90 analyze  --filter-tags=inexistant
 
 
 echo "Test 8: --share-email allowed max 3 times"
-isitfit --share-email=abc --share-email=fdas --share-email=fsf --share-email=fdasf cost analyze --ndays=90 || echo "expected to fail"
+isitfit --share-email=abc --share-email=fdas --share-email=fsf --share-email=fdasf cost --ndays=90 analyze || echo "expected to fail"
 
 
 echo "Test 9: --share-email ok"
-AWS_PROFILE=shadi_shadi isitfit --share-email=shadi@autofitcloud.com cost analyze --ndays=90
+AWS_PROFILE=shadi_shadi isitfit --share-email=shadi@autofitcloud.com cost --ndays=90 analyze
 
 
 # restore the original UID